Refactor handlers and rename classes
