+++ root/app/Models/Account.php
@@ -2,7 +2,7 @@
 namespace App\Models;
 
 use Exception;
-use App\Models\DatabaseHandler;
+use App\Models\Database;
 use App\Core\ErrorHandler;
 
 /**
@@ -25,7 +25,7 @@ class AccountHandler // @phpcs:disable PSR1.Classes.ClassDeclaration.MissingName
     public static function getAllAccounts(): array
     {
         try {
-            $db = new DatabaseHandler();
+            $db = new Database();
             $db->query("SELECT * FROM accounts");
             return $db->resultSet();
         } catch (Exception $e) {
@@ -43,7 +43,7 @@ class AccountHandler // @phpcs:disable PSR1.Classes.ClassDeclaration.MissingName
     public static function getAllUserAccts(string $username): array
     {
         try {
-            $db = new DatabaseHandler();
+            $db = new Database();
             $db->query("SELECT account FROM accounts WHERE username = :username");
             $db->bind(':username', $username);
             return $db->resultSet();
@@ -63,7 +63,7 @@ class AccountHandler // @phpcs:disable PSR1.Classes.ClassDeclaration.MissingName
     public static function accountExists(string $accountOwner, string $accountName): bool
     {
         try {
-            $db = new DatabaseHandler();
+            $db = new Database();
             $db->query("SELECT 1 FROM accounts WHERE username = :accountOwner AND account = :accountName LIMIT 1");
             $db->bind(':accountOwner', $accountOwner);
             $db->bind(':accountName', $accountName);
@@ -84,7 +84,7 @@ class AccountHandler // @phpcs:disable PSR1.Classes.ClassDeclaration.MissingName
     public static function getAcctInfo(string $username, string $account): mixed
     {
         try {
-            $db = new DatabaseHandler();
+            $db = new Database();
             $db->query("SELECT * FROM accounts WHERE username = :username AND account = :account");
             $db->bind(':username', $username);
             $db->bind(':account', $account);
@@ -105,7 +105,7 @@ class AccountHandler // @phpcs:disable PSR1.Classes.ClassDeclaration.MissingName
     public static function getAccountLink(string $username, string $account): string
     {
         try {
-            $db = new DatabaseHandler();
+            $db = new Database();
             $db->query("SELECT link FROM accounts WHERE username = :username AND account = :account");
             $db->bind(':username', $username);
             $db->bind(':account', $account);
@@ -132,7 +132,7 @@ class AccountHandler // @phpcs:disable PSR1.Classes.ClassDeclaration.MissingName
      */
     public static function updateAccount(string $accountOwner, string $accountName, string $prompt, string $platform, int $hashtags, string $link, string $cron, string $days): bool
     {
-        $db = new DatabaseHandler();
+        $db = new Database();
         $db->beginTransaction();
         try {
             $db->query("UPDATE accounts SET prompt = :prompt, platform = :platform, hashtags = :hashtags, link = :link, cron = :cron, days = :days WHERE username = :accountOwner AND account = :accountName");
@@ -169,7 +169,7 @@ class AccountHandler // @phpcs:disable PSR1.Classes.ClassDeclaration.MissingName
      */
     public static function createAccount(string $accountOwner, string $accountName, string $prompt, string $platform, int $hashtags, string $link, string $cron, string $days): bool
     {
-        $db = new DatabaseHandler();
+        $db = new Database();
         $db->beginTransaction();
         try {
             $db->query("INSERT INTO accounts (username, account, prompt, platform, hashtags, link, cron, days) VALUES (:accountOwner, :accountName, :prompt, :platform, :hashtags, :link, :cron, :days)");
@@ -200,7 +200,7 @@ class AccountHandler // @phpcs:disable PSR1.Classes.ClassDeclaration.MissingName
      */
     public static function deleteAccount(string $accountOwner, string $accountName): bool
     {
-        $db = new DatabaseHandler();
+        $db = new Database();
         $db->beginTransaction();
         try {
             $db->query("DELETE FROM status_updates WHERE username = :accountOwner AND account = :accountName");
