+++ root/app/Models/Database.php
@@ -10,12 +10,12 @@ use Exception;
  * Author: Vontainment
  * URL: https://vontainment.com/
  * Version: 2.0.0
- * File: DatabaseHandler.php
+ * File: Database.php
  * Description: Manages database connection and queries using PDO.
  * License: MIT
  */
 
-class DatabaseHandler // @phpcs:disable PSR1.Classes.ClassDeclaration.MissingNamespace
+class Database // @phpcs:disable PSR1.Classes.ClassDeclaration.MissingNamespace
 {
     private static ?PDO $dbh = null;
     private static ?int $lastUsedTime = null; // Tracks the last time the connection was used
