+++ root/app/Models/User.php
@@ -2,7 +2,7 @@
 namespace App\Models;
 
 use Exception;
-use App\Models\DatabaseHandler;
+use App\Models\Database;
 use App\Core\ErrorHandler;
 
 /**
@@ -25,7 +25,7 @@ class UserHandler // @phpcs:disable PSR1.Classes.ClassDeclaration.MissingNamespa
     public static function getAllUsers(): array
     {
         try {
-            $db = new DatabaseHandler();
+            $db = new Database();
             $db->query("SELECT * FROM users");
             return $db->resultSet();
         } catch (Exception $e) {
@@ -43,7 +43,7 @@ class UserHandler // @phpcs:disable PSR1.Classes.ClassDeclaration.MissingNamespa
     public static function userExists(string $username): mixed
     {
         try {
-            $db = new DatabaseHandler();
+            $db = new Database();
             $db->query("SELECT * FROM users WHERE username = :username");
             $db->bind(':username', $username);
             return $db->single();
@@ -69,7 +69,7 @@ class UserHandler // @phpcs:disable PSR1.Classes.ClassDeclaration.MissingNamespa
     public static function updateUser(string $username, string $password, int $totalAccounts, int $maxApiCalls, int $usedApiCalls, string $expires, int $admin, bool $isUpdate): bool
     {
         ErrorHandler::logMessage("updateUser called with username: $username, isUpdate: $isUpdate", 'info');
-        $db = new DatabaseHandler();
+        $db = new Database();
         $db->beginTransaction();
         try {
             if ($isUpdate) {
@@ -102,7 +102,7 @@ class UserHandler // @phpcs:disable PSR1.Classes.ClassDeclaration.MissingNamespa
      */
     public static function deleteUser(string $username): bool
     {
-        $db = new DatabaseHandler();
+        $db = new Database();
         $db->beginTransaction();
         try {
             $db->query("DELETE FROM users WHERE username = :username");
@@ -136,7 +136,7 @@ class UserHandler // @phpcs:disable PSR1.Classes.ClassDeclaration.MissingNamespa
     public static function updatePassword(string $username, string $hashedPassword): bool
     {
         try {
-            $db = new DatabaseHandler();
+            $db = new Database();
             $db->query("UPDATE users SET password = :password WHERE username = :username");
             $db->bind(':username', $username);
             $db->bind(':password', $hashedPassword);
@@ -157,7 +157,7 @@ class UserHandler // @phpcs:disable PSR1.Classes.ClassDeclaration.MissingNamespa
     public static function getUserInfo(string $username): mixed
     {
         try {
-            $db = new DatabaseHandler();
+            $db = new Database();
             $db->query("SELECT * FROM users WHERE username = :username");
             $db->bind(':username', $username);
             return $db->single();
@@ -176,7 +176,7 @@ class UserHandler // @phpcs:disable PSR1.Classes.ClassDeclaration.MissingNamespa
     public static function getAllUserAccts(string $username): array
     {
         try {
-            $db = new DatabaseHandler();
+            $db = new Database();
             $db->query("SELECT * FROM accounts WHERE username = :username");
             $db->bind(':username', $username);
             return $db->resultSet();
@@ -196,7 +196,7 @@ class UserHandler // @phpcs:disable PSR1.Classes.ClassDeclaration.MissingNamespa
     public static function updateUsedApiCalls(string $username, int $usedApiCalls): bool
     {
         try {
-            $db = new DatabaseHandler();
+            $db = new Database();
             $db->query("UPDATE users SET used_api_calls = :used_api_calls WHERE username = :username");
             $db->bind(':used_api_calls', $usedApiCalls);
             $db->bind(':username', $username);
@@ -218,7 +218,7 @@ class UserHandler // @phpcs:disable PSR1.Classes.ClassDeclaration.MissingNamespa
     public static function updateMaxApiCalls(string $username, int $maxApiCalls): bool
     {
         try {
-            $db = new DatabaseHandler();
+            $db = new Database();
             $db->query("UPDATE users SET max_api_calls = :max_api_calls WHERE username = :username");
             $db->bind(':max_api_calls', $maxApiCalls);
             $db->bind(':username', $username);
@@ -238,7 +238,7 @@ class UserHandler // @phpcs:disable PSR1.Classes.ClassDeclaration.MissingNamespa
     public static function resetAllApiUsage(): bool
     {
         try {
-            $db = new DatabaseHandler();
+            $db = new Database();
             $db->query("UPDATE users SET used_api_calls = 0");
             $db->execute();
             return true;
@@ -261,7 +261,7 @@ class UserHandler // @phpcs:disable PSR1.Classes.ClassDeclaration.MissingNamespa
     public static function updateProfile(string $username, string $who, string $where, string $what, string $goal): bool
     {
         try {
-            $db = new DatabaseHandler();
+            $db = new Database();
             $db->query("UPDATE users SET who = :who, `where` = :where, what = :what, goal = :goal WHERE username = :username");
             $db->bind(':username', $username);
             $db->bind(':who', $who);
